# adan-platform/contracts/openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ADAN API",
    "description": "Sistema Operativo Empresarial - Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/kg/nodes": {
      "post": {
        "tags": [
          "knowledge-graph"
        ],
        "summary": "Create Node",
        "operationId": "create_node_kg_nodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/kg/edges": {
      "post": {
        "tags": [
          "knowledge-graph"
        ],
        "summary": "Create Edge",
        "operationId": "create_edge_kg_edges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEdgeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/kg/nodes/{node_id}": {
      "get": {
        "tags": [
          "knowledge-graph"
        ],
        "summary": "Get Node",
        "operationId": "get_node_kg_nodes__node_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kg/proyectos/{proyecto_id}/nodes": {
      "get": {
        "tags": [
          "knowledge-graph"
        ],
        "summary": "List Nodes By Proyecto",
        "operationId": "list_nodes_by_proyecto_kg_proyectos__proyecto_id__nodes_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proyecto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proyecto Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NodeResponse"
                  },
                  "title": "Response List Nodes By Proyecto Kg Proyectos  Proyecto Id  Nodes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kg/nodes/{node_id}/traverse": {
      "get": {
        "tags": [
          "knowledge-graph"
        ],
        "summary": "Traverse From Node",
        "operationId": "traverse_from_node_kg_nodes__node_id__traverse_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "max_depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2,
              "title": "Max Depth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraverseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/kg/query": {
      "get": {
        "tags": [
          "knowledge-graph"
        ],
        "summary": "Hybrid Query",
        "description": "Combina busqueda semantica (pgvector cosine_distance) + traversal del grafo desde\nlos mejores resultados semanticos - misma logica que ai/memory/hybrid_query.py,\nreimplementada aqui sobre los modelos espejo del backend (sin import cruzado,\nPlan Maestro SS3.2). Ver docstring de ese modulo para la justificacion completa.",
        "operationId": "hybrid_query_kg_query_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "proyecto_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proyecto Id"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Top K"
            }
          },
          {
            "name": "graph_depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Graph Depth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HybridResultResponse"
                  },
                  "title": "Response Hybrid Query Kg Query Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateEdgeRequest": {
        "properties": {
          "origen_id": {
            "type": "string",
            "title": "Origen Id"
          },
          "destino_id": {
            "type": "string",
            "title": "Destino Id"
          },
          "tipo_relacion": {
            "type": "string",
            "title": "Tipo Relacion"
          },
          "atributos": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atributos"
          }
        },
        "type": "object",
        "required": [
          "origen_id",
          "destino_id",
          "tipo_relacion"
        ],
        "title": "CreateEdgeRequest"
      },
      "CreateNodeRequest": {
        "properties": {
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "proyecto_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proyecto Id"
          },
          "atributos": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atributos"
          }
        },
        "type": "object",
        "required": [
          "tipo",
          "nombre"
        ],
        "title": "CreateNodeRequest"
      },
      "EdgeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "origen_id": {
            "type": "string",
            "title": "Origen Id"
          },
          "destino_id": {
            "type": "string",
            "title": "Destino Id"
          },
          "tipo_relacion": {
            "type": "string",
            "title": "Tipo Relacion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "origen_id",
          "destino_id",
          "tipo_relacion"
        ],
        "title": "EdgeResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HybridResultResponse": {
        "properties": {
          "contenido": {
            "type": "string",
            "title": "Contenido"
          },
          "origen": {
            "type": "string",
            "title": "Origen"
          },
          "distance": {
            "type": "number",
            "title": "Distance"
          },
          "nodo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nodo Id"
          }
        },
        "type": "object",
        "required": [
          "contenido",
          "origen",
          "distance",
          "nodo_id"
        ],
        "title": "HybridResultResponse"
      },
      "NodeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "proyecto_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proyecto Id"
          },
          "atributos": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atributos"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tipo",
          "nombre",
          "proyecto_id",
          "atributos"
        ],
        "title": "NodeResponse"
      },
...
      "TraverseResponse": {
        "properties": {
          "nodos": {
            "items": {
              "$ref": "#/components/schemas/NodeResponse"
            },
            "type": "array",
            "title": "Nodos"
          },
          "aristas": {
            "items": {
              "$ref": "#/components/schemas/EdgeResponse"
            },
            "type": "array",
            "title": "Aristas"
          }
        },
        "type": "object",
        "required": [
          "nodos",
          "aristas"
        ],
        "title": "TraverseResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}